[astro-purgecss] Fix race-condition in calls to replaceValueInFile - fix #660
Ensure that all CSS file references needing to be updated are done before saving HTML files.

The process has been refactored in order to:

* read HTML files only once
* replace all CSS file references which need to be changed (for loop)
* write updated HTML file to disk.

As a side-effect, build performance should also be improved (previous version was reading and writing file each time a CSS file reference had to be updated).